Fix Multicast Delegate invokation order
Accordingly to mentor's comment:
I was debugging the app to find some bugs and found a following issue. Take a look on the console
output:
(lldb) po delegates.allObjects.reversed()
▿ ReversedCollection<Array<AnyObject>>
▿ _base : 5 elements
    ▿ 0 : <Project1.SolarSystemDetailedViewController: 0x7f9dab91b820>
    ▿ 1 : <Project1.UniversesViewController: 0x7f9dab9096b0>
    ▿ 2 : <Storage: 0x600001180d00>
    ▿ 3 : <Project1.GalaxyViewController: 0x7f9da9c2d6b0>
    ▿ 4 : <Project1.GalaxiesViewController: 0x7f9da9f14ad0>
The code below calls delegates in a sequential order, and the general logic of data reload
is built around an idea of the datasource (Storage) being updated first, and the UI
(any UIViewController that extends BaseViewController) next.
However the console's output tells us that some screens will not reload any data because
they are called before the storage. Of course eventually they will be called on the next tick,
so perhaps a user wouldn't notice anything so this isn't a crictical issue, rather an architectural one.